# a7_openapi.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: 'https://virtserver.swaggerhub.com/grupo_lbaw_2425/projeto_lbaw_2425_openapi/1.0.0'
info:
  version: 1.0.0
  title: FEUPshare-api
  description: The Web API for FEUPshare
...
paths:
...
  /api/tags:
    get:
      summary: "Search Tags"
      description: "Tag search results returned successfully. Access: USR, MOD, ADM"
      operationId: searchTags
      tags:
        - "M06: Search"
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Tag search results returned successfully. Access: USR, MOD, ADM"
...
  /tags:
    get:
      summary: "List Tags"
      description: "List of tags retrieved successfully. Access: PUB"
      operationId: listTags
      tags:
        - "M08: Static Pages"
      responses:
        '200':
          description: "List of tags retrieved successfully. Access: PUB"
    post:
      summary: "Create New Tag"
      description: "Tag created successfully. Access: PUB"
      operationId: createTag
      tags:
        - "M08: Static Pages"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        '201':
          description: "Tag created successfully. Access: PUB"
...